add list update
add list update

+++ web/WebContent/jsp/passwordConfirm.jsp
@@ -19,13 +19,18 @@
 		
 		if(dbPass!=null && cp.equals(dbPass.trim())){
 		//비밀번호가 맞을경우
-		%>
-		<script>
-		alert("정말 삭제하시겠습니까?");
-		</script>
-		<%
+			if(job.equals("del")){
+			%>
+			<script>
+			alert("정말 삭제하시겠습니까?");
+			</script>
+			<%
 			AppleDao.deleteBBS(no);
 			response.sendRedirect("list.jsp");
+			
+			}else if(job.equals("modify")){
+				response.sendRedirect("edit.jsp?mode=modify&no="+no+"&page="+pageData);
+			}
 		}else{
 		//비밀번호가 틀릴경우
 			response.sendRedirect("password.jsp?no="+no+"&job="+job+"&page="+pageData);
